Create Missing numbers
https://www.hackerrank.com/challenges/missing-numbers/problem


Given two arrays of integers, find which elements in the second array are missing from the first array.
Example
203 204 205 206 207 208 203 204 205 206
203 204 204 205 206 207 205 208 203 206 205 206 204
The  array is the orginal list. The numbers missing are .
204 205 206
Notes

If a number occurs multiple times in the lists, you must ensure that the frequency of that number in both lists is the same. If that is not the case, then it is also a missing number.
Return the missing numbers sorted ascending.
Only include a missing number once, even if it is missing multiple times.
The difference between the maximum and minimum numbers in the original list is less than or equal to .